+++ css/comment.css
@@ -1,0 +1,160 @@
+* {
+  font-family: "SUIT", sans-serif;
+  font-weight: 500;
+  line-height: 160%;
+}
+
+.comment {
+  margin: 5vh 10vh 5vh 10vh;
+}
+
+.commentBody {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, 400px);
+  gap: 50px;
+  justify-content: center;
+}
+
+.commentBox {
+  justify-self: center;
+  display: flex;
+  flex-direction: column;
+  height: 120px;
+  width: 350px;
+  border-radius: 10px;
+  border: 1px solid #ff7211;
+  padding: 15px;
+  position: relative;
+}
+
+.commentBox p {
+  position: absolute;
+  margin: 0;
+  padding: 0;
+}
+
+.commentText {
+  top: 40px;
+  font-size: 20px;
+  line-height: 1.3;
+}
+
+.commentTime {
+  font-size: 12px;
+  bottom: 10px;
+  color: #616161;
+}
+
+.commentUser {
+  font-size: 15px;
+  bottom: 23px;
+  color: #616161;
+}
+
+.commentStar {
+  color: #ff7211;
+  font-size: 20px;
+  top: 10px;
+}
+
+/* 수정, 삭제 버튼 */
+.commentBox button {
+  position: absolute;
+  bottom: 15px;
+  border-radius: 10px;
+  border: none;
+  width: 45px;
+  color: white;
+  font-weight: 200;
+}
+
+.commentBox > .del {
+  right: 15px;
+  background-color: #42abae;
+}
+
+.commentBox > .edit {
+  right: 65px;
+  background-color: #f18910;
+}
+
+/* 댓글 입력 부분 */
+.writeBox {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+#writeBtn {
+  border-radius: 20px;
+  width: 150px;
+  height: 40px;
+  background-color: #0da9b3;
+  border: none;
+  color: white;
+  margin-bottom: 50px;
+}
+
+.commnetWrite {
+  display: flex;
+  flex-direction: row;
+  margin-bottom: 70px;
+  align-items: center;
+}
+
+.writeBox input {
+  border-radius: 10px;
+  border: 1px solid #0da9b3;
+  background-color: transparent;
+  text-align: center;
+}
+
+.writeBox input:focus {
+  outline: none;
+}
+
+#inputUser,
+#inputPw {
+  width: 150px;
+  height: 27px;
+  margin: 3px 0;
+}
+
+#inputComment {
+  width: 500px;
+  height: 65px;
+  margin: 0 20px;
+  flex-wrap: wrap;
+}
+
+#commnetBtn {
+  width: 120px;
+  height: 35px;
+  border-radius: 10px;
+  border: none;
+  background-color: #0da9b3;
+  color: white;
+}
+
+.selectStar {
+  border-radius: 10px;
+  height: 28px;
+  width: 120px;
+  margin-bottom: 5px;
+  background-color: #0da9b3;
+  border: none;
+  color: white;
+  appearance: none;
+  padding-left: 10px;
+  position: relative;
+}
+
+.starArrow {
+  position: absolute;
+  right: 10px;
+  top: -1%;
+}
+
+.selectStar:focus {
+  outline: none;
+}
