+++ css/detail.css
@@ -1,0 +1,53 @@
+.posterDiv {
+  height: 1000px;
+  display: grid;
+  transform: translate(-10%, 0%);
+  justify-content: center;
+  grid-template-columns: repeat(70, 2em);
+  grid-template-rows: repeat(50, 1em);
+}
+#item:nth-child(1) {
+  grid-column: 30;
+  grid-row: 10;
+  transform: translate(30px, 0px);
+}
+#item:nth-child(2) {
+  grid-column: 48 / 60;
+  grid-row: 12;
+  font-weight: bolder;
+  font-size: 20px;
+}
+#item:nth-child(3) {
+  grid-row: 16;
+  grid-column: 48 / 55;
+}
+#item:nth-child(4) {
+  grid-column: 48/ 53;
+  grid-row: 20;
+}
+#item:nth-child(5) {
+  grid-column: 53 / 55;
+  grid-row: 20;
+  width: 130px;
+  margin-right: 50px;
+}
+#item:nth-child(6) {
+  grid-column: 57/ 65;
+  grid-row: 20;
+}
+#item:nth-child(7) {
+  grid-column: 48 / 65;
+  grid-row: 20;
+  margin-top: 80px;
+}
+.actorContainer {
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: row;
+  justify-content: space-evenly;
+  list-style: none;
+}
+
+.actorImg {
+  width: 300px;
+}
