+++ detailpage.html
@@ -3,71 +3,15 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./style.css" />
+    <link rel="stylesheet" href="./css/style.css" />
+    <link rel="stylesheet" href="./css/comment.css" />
+    <link rel="stylesheet" href="./css/detail.css" />
     <link href="https://cdn.jsdelivr.net/gh/sun-typeface/SUIT@2/fonts/static/woff2/SUIT.css" rel="stylesheet" />
     <link
       rel="stylesheet"
       href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
     />
-    <link rel="stylesheet" href="./style.css" />
-    <script src="./movies.js" defer></script>
-
     <title>Movie Details</title>
-    <style>
-      .posterDiv {
-        height: 1000px;
-        display: grid;
-        transform: translate(-10%, 0%);
-        justify-content: center;
-        grid-template-columns: repeat(70, 2em);
-        grid-template-rows: repeat(50, 1em);
-      }
-      #item:nth-child(1) {
-        grid-column: 30;
-        grid-row: 10;
-        transform: translate(30px, 0px);
-      }
-      #item:nth-child(2) {
-        grid-column: 48 / 60;
-        grid-row: 12;
-        font-weight: bolder;
-        font-size: 20px;
-      }
-      #item:nth-child(3) {
-        grid-row: 16;
-        grid-column: 48 / 55;
-      }
-      #item:nth-child(4) {
-        grid-column: 48/ 53;
-        grid-row: 20;
-      }
-      #item:nth-child(5) {
-        grid-column: 53 / 55;
-        grid-row: 20;
-        width: 130px;
-        margin-right: 50px;
-      }
-      #item:nth-child(6) {
-        grid-column: 57/ 65;
-        grid-row: 20;
-      }
-      #item:nth-child(7) {
-        grid-column: 48 / 65;
-        grid-row: 20;
-        margin-top: 80px;
-      }
-      .actorContainer {
-        display: flex;
-        flex-wrap: wrap;
-        flex-direction: row;
-        justify-content: space-evenly;
-        list-style: none;
-      }
-
-      .actorImg {
-        width: 300px;
-      }
-    </style>
   </head>
   <body>
     <header id="header">
@@ -98,10 +42,11 @@
           stroke-linecap="round"
         />
       </svg>
+
       <div class="search">
         <input type="text" id="search-input" onkeyup="enterkey()" />
         <button id="search-button" class="search-button">
-          <svg width="40px" height="auto" viewBox="0 0 58 57" fill="none" xmlns="http://www.w3.org/2000/svg">
+          <svg width="40px" viewBox="0 0 58 57" fill="none" xmlns="http://www.w3.org/2000/svg">
             <path
               d="M30.4255 3.70679C20.0133 3.70679 7.95417 14.9149 4.10072 24.0799C-0.814613 35.7704 12.0113 45.9603 21.269 49.7179C31.6028 53.9122 41.2318 49.3278 47.1932 40.2753C54.214 29.6141 51.7222 15.1636 40.8409 8.51392C36.6743 5.96764 29.9455 6.79708 25.275 6.79708"
               stroke="#fbbc14"
@@ -122,5 +67,75 @@
     <div class="detaillWrap"></div>
 
     <ul class="actorContainer"></ul>
+
+    <div class="comment">
+      <div class="writeBox">
+        <div>
+          <button id="writeBtn">리뷰 남기기</button>
+        </div>
+        <div class="commnetWrite" id="commnetWrite">
+          <div style="display: flex; flex-direction: column">
+            <input id="inputUser" type="text" placeholder="닉네임" />
+            <input id="inputPw" type="password" placeholder="비밀번호" />
+          </div>
+          <input id="inputComment" type="text" placeholder="한 줄 리뷰를 작성해주세요! (최대 40자)" maxlength="40" />
+          <div style="display: flex; flex-direction: column; position: relative">
+            <select class="selectStar" id="star">
+              <option>별점</option>
+              <option value="★">⭐</option>
+              <option value="★★">⭐⭐</option>
+              <option value="★★★">⭐⭐⭐</option>
+              <option value="★★★★">⭐⭐⭐⭐</option>
+              <option value="★★★★★">⭐⭐⭐⭐⭐</option>
+            </select>
+            <div class="starArrow">
+              <svg fill="white" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" style="height: 25px">
+                <path
+                  d="M137.4 374.6c12.5 12.5 32.8 12.5 45.3 0l128-128c9.2-9.2 11.9-22.9 6.9-34.9s-16.6-19.8-29.6-19.8L32 192c-12.9 0-24.6 7.8-29.6 19.8s-2.2 25.7 6.9 34.9l128 128z"
+                />
+              </svg>
+            </div>
+            <button id="commnetBtn">등록</button>
+          </div>
+        </div>
+      </div>
+      <div class="commentBody" id="commentBody">
+        <div class="commentBox" id="commentBox">
+          <p class="commentStar" id="commentStar">★★★★★</p>
+          <p class="commentUser" id="commentUser">닉네임</p>
+          <p class="commentText" id="commentText">댓글내용블라블라</p>
+          <p class="commentTime" id="commentTime">24.08.01</p>
+          <button class="edit">수정</button>
+          <button class="del">삭제</button>
+        </div>
+        <div class="commentBox" id="commentBox">
+          <p class="commentStar" id="commentStar">★★★★★</p>
+          <p class="commentUser" id="commentUser">닉네임</p>
+          <p class="commentText" id="commentText">댓글내용블라블라</p>
+          <p class="commentTime" id="commentTime">24.08.01</p>
+          <button class="edit" id="commentEdit">수정</button>
+          <button class="del" id="commentDel">삭제</button>
+        </div>
+        <div class="commentBox" id="commentBox">
+          <p class="commentStar" id="commentStar">★★★★★</p>
+          <p class="commentUser" id="commentUser">닉네임</p>
+          <p class="commentText" id="commentText">댓글내용블라블라</p>
+          <p class="commentTime" id="commentTime">24.08.01</p>
+          <button class="edit" id="commentEdit">수정</button>
+          <button class="del" id="commentDel">삭제</button>
+        </div>
+        <div class="commentBox" id="commentBox">
+          <p class="commentStar" id="commentStar">★★★★★</p>
+          <p class="commentUser" id="commentUser">닉네임</p>
+          <p class="commentText" id="commentText">댓글내용블라블라</p>
+          <p class="commentTime" id="commentTime">24.08.01</p>
+          <button class="edit" id="commentEdit">수정</button>
+          <button class="del" id="commentDel">삭제</button>
+        </div>
+      </div>
+    </div>
+    <button id="top-btn" onclick="toTop()">▲<br />Top</button>
+    <script type="module" src="./script/detail.js"></script>
+    <script src="./script/comment.js"></script>
   </body>
 </html>
