+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Movie Box</title>
     <link href="https://cdn.jsdelivr.net/gh/sun-typeface/SUIT@2/fonts/static/woff2/SUIT.css" rel="stylesheet" />
-    <link rel="stylesheet" href="./style.css" />
+    <link rel="stylesheet" href="./css/style.css" />
     <link
       rel="stylesheet"
       href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
@@ -37,7 +37,7 @@
     <div class="search">
       <input type="text" id="search-input" onkeyup="enterkey()" />
       <button id="search-button" class="search-button">
-        <svg width="40px" height="auto" viewBox="0 0 58 57" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <svg width="40px" viewBox="0 0 58 57" fill="none" xmlns="http://www.w3.org/2000/svg">
           <path
             d="M30.4255 3.70679C20.0133 3.70679 7.95417 14.9149 4.10072 24.0799C-0.814613 35.7704 12.0113 45.9603 21.269 49.7179C31.6028 53.9122 41.2318 49.3278 47.1932 40.2753C54.214 29.6141 51.7222 15.1636 40.8409 8.51392C36.6743 5.96764 29.9455 6.79708 25.275 6.79708"
             stroke="#fbbc14"
@@ -56,8 +56,8 @@
   </header>
 
   <body>
-    <!-- 개봉 예정작 -->
     <div class="movieBody">
+      <!-- 개봉 예정작 -->
       <div class="slideBox">
         <p class="categoryTitle">개봉 예정작</p>
         <!-- 슬라이더 부분 -->
@@ -67,12 +67,25 @@
 
     <!-- 현재 상영작 -->
     <div class="slideBox">
-      <p class="categoryTitle">현재 상영작</p>
+      <p class="categoryTitle">절찬리 상영중!</p>
       <!-- 슬라이더 부분 -->
       <div id="nowContainer" class="Container"></div>
     </div>
+    <!-- 현재 상영작 -->
+    <div class="slideBox">
+      <p class="categoryTitle">사람들이 가장 많이 본 영화</p>
+      <!-- 슬라이더 부분 -->
+      <div id="popularContainer" class="Container"></div>
+    </div>
+
+    <!-- 현재 상영작 -->
+    <div class="slideBox">
+      <p class="categoryTitle">안보면 후회하는 명작</p>
+      <!-- 슬라이더 부분 -->
+      <div id="topContainer" class="Container"></div>
+    </div>
 
     <button id="top-btn" onclick="toTop()">▲<br />Top</button>
-    <script src="./movies.js"></script>
+    <script type="module" src="./script/movies.js"></script>
   </body>
 </html>
